backend/src/requirements.txt
fastapi==0.129.2
peft==0.18.1
pillow==12.1.1
pip==26.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.22
PyYAML==6.0.3
SQLAlchemy==2.0.46
torch==2.10.0
transformers==4.57.6
uvicorn==0.41.0
